RavenDB-27248 Merge the CDC date/time tests into a single Issues/RavenDB_27248 test file
